% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get point region.R
\name{get_point_region_from_shape}
\alias{get_point_region_from_shape}
\title{Is a latitude and longitude point in a matrix}
\usage{
get_point_region_from_shape(shp, region_name, lon, lat)
}
\arguments{
\item{shp}{shape (S4)}

\item{region_name}{String for region you want to check}

\item{lon}{Longitude (can be a vector)}

\item{lat}{Latitude (can be a vector)}
}
\description{
This function takes a shape, the region_name of the region you want to check eg. "NAME_4", vector of lon
vector of lat values, and returns the regions the lat/lon points are in
}
