% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what subdivision is point in.R
\name{what_subdivision_is_point_in}
\alias{what_subdivision_is_point_in}
\title{Is a latitude and longitude point in a matrix}
\usage{
what_subdivision_is_point_in(shp, colname, lon, lat)
}
\arguments{
\item{shp}{shape (S4)}

\item{colname}{Column name of the layer}

\item{lon}{Longitude (can be a vector)}

\item{lat}{Latitude (can be a vector)}
}
\description{
This function takes a shape, the column name of the regions in the shape, and lon and lat
}
\details{
If the map show that your point should be inside the shapefile, but is returning false, it's probably because
there is no polygon over that area. Easiest way to fix this is to try a different shapefile of the same thing. Note: This
function probably won't work as you expect.
}
